Initial Commit
Create a framework for auto injecting dependencies via a configuration file

The components make up a set of services that are used to wire together a system for managing jobs.
Pieces of the system can be replaced with technology specific functionality.
IDatastore for blob storage or file storage
IDocumentStore for document storage around the content in the file storage
IJobManager manages jobs how you want
IWorkerManager will handle doling out work
IWorker handles the work itself

Decorate your class with an attribute depending on what the class is specified to do and it'll inject it and all of its interfaces into the dependency chain